ui_web/requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
reportlab>=4.0.0
